Make omnigot_v2 the default one
