+++ meta_blocks/datasets/__init__.py
@@ -1,4 +1,4 @@
-from meta_blocks.datasets import omniglot_v2 as omniglot
+from meta_blocks.datasets import omniglot
 from meta_blocks.datasets.base import *
