+++ meta_blocks/datasets/omniglot.py
@@ -1,51 +1,39 @@
-"""
-Loading and augmenting the Omniglot dataset.
-
-To use these APIs, you should prepare a directory that
-contains all of the alphabets from both images_background
-and images_evaluation.
-"""
+"""The Omniglot dataset."""
 
 import glob
 import logging
 import os
 import random
+from typing import Any, Callable, List, Optional, Tuple
 
+import numpy as np
 import tensorflow.compat.v1 as tf
+from PIL import Image
 
-from meta_blocks.datasets import base, utils
+from meta_blocks.datasets import base
 
 logger = logging.getLogger(__name__)
 
-__all__ = ["OmniglotCategory", "OmniglotDataset", "OmniglotMetaDataset"]
+# Transition to V2 will happen in stages.
+tf.disable_v2_behavior()
+tf.enable_resource_variables()
+
+__all__ = [
+    "OmniglotCategory",
+    "OmniglotDataPool",
+    "OmniglotDataset",
+    "OmniglotMetaDataset",
+]
 
 
-def read_dataset(
+def get_categories(
     data_dir,
     num_train_categories=1000,
     num_valid_categories=200,
     num_test_categories=463,
+    shuffle=True,
 ):
-    """Iterate over the characters in a data directory.
-    The dataset is unaugmented and not split up into training and test sets.
-
-    Parameters
-    ----------
-    data_dir : str
-        A directory of alphabet directories.
-
-    num_train_categories : int, optional (default=1000)
-
-    num_valid_categories : int, optional (default=200)
-
-    num_test_categories : int, optional (default=463)
-
-    Returns
-    -------
-    train_test_val_set : dict
-        An iterable over Characters.
-    """
-
+    logger.info("Reading Omniglot categories...")
     categories = []
     for alphabet_name in sorted(os.listdir(data_dir)):
         alphabet_dir = os.path.join(data_dir, alphabet_name)
@@ -56,150 +44,166 @@ def read_dataset(
                 continue
             if not name.startswith("character"):
                 continue
-            c = OmniglotCategory(data_dir, os.path.join(alphabet_name, name))
-            categories.append(c)
-    random.shuffle(categories)
-    return {
-        "train": categories[:num_train_categories],
-        "valid": categories[num_train_categories:][:num_valid_categories],
-        "test": categories[num_train_categories:][-num_test_categories:],
-    }
+            char_name = f"{alphabet_name}_{name}"
+            char_dir = os.path.join(data_dir, alphabet_name, name)
+            categories.append(OmniglotCategory(char_dir, char_name))
+    if shuffle:
+        random.shuffle(categories)
+    categories_dict = {"train": categories[:num_train_categories]}
+    if num_valid_categories is not None:
+        categories_dict["valid"] = categories[num_train_categories:][
+            :num_valid_categories
+        ]
+    if num_test_categories is not None:
+        categories_dict["test"] = categories[num_train_categories:][
+            -num_test_categories:
+        ]
+    return categories_dict
 
 
 class OmniglotCategory(base.Category):
-    """A single character class.
-
-    Parameters
-    ----------
-    data_dir : str
-        The data directory.
-
-    name : str
-        Name of the category.
-
-    maybe_gen_tfrecords : bool, optional (default=False)
-        Whether to generate TF records. Deprecated.
-    """
-
-    def __init__(self, data_dir, name, maybe_gen_tfrecords=False):
-
-        super(OmniglotCategory, self).__init__(data_dir, name)
-        self._maybe_gen_tfrecords = maybe_gen_tfrecords
-
-    def maybe_generate_tfrecords(self, tfrecord_path):
-        """Generates TFRecords for the corresponding data."""
-        # Get file paths and determine dataset size.
-        filepaths = glob.glob(os.path.join(self.data_dir, self.name, "*.png"))
-        self._dataset_size = len(filepaths)
-
-        # Generate TFRecords, if necessary.
-        if os.path.exists(tfrecord_path):
-            logger.debug(
-                f"Category {self.name} in {self.data_dir} "
-                f"is already preprocessed. Skipping..."
-            )
-            return
-        logger.debug(
-            f"Generating TFRecords for category {self.name} " f"in {self.data_dir}..."
-        )
-
-        # Read images and construct features.
-        features = []
-        for fpath in sorted(filepaths):
-            with open(fpath, "rb") as fp:
-                feature = {"image_raw": utils.bytes_feature_list([fp.read()])}
-                features.append(feature)
-
-        # Write features to a TFRecords file.
-        with tf.python_io.TFRecordWriter(tfrecord_path) as writer:
-            for feature in features:
-                example = tf.train.Example(features=tf.train.Features(feature=feature))
-                writer.write(example.SerializeToString())
-
-        logger.info("Done.")
-
-    def build(
-        self,
-        size=None,
-        shuffle=False,
-        num_parallel_reads=16,
-        shuffle_buffer_size=100,
-        seed=None,
-        **kwargs,
-    ):
-        """Builds tf.data.Dataset for the underlying data resources."""
-        tfrecord_path = os.path.join(self.data_dir, f"{self.name}.tfrecords")
-
-        # Generate TFRecords.
-        if self._maybe_gen_tfrecords:
-            self.maybe_generate_tfrecords(tfrecord_path)
-        else:
-            self._dataset_size = 20
-
-        # Build the tf.data.Dataset.
-        self._dataset = tf.data.TFRecordDataset(
-            tfrecord_path, num_parallel_reads=num_parallel_reads
-        )
-        if shuffle:
-            self._dataset = self._dataset.shuffle(shuffle_buffer_size, seed=seed)
-        self._size = (
-            self._dataset_size if size is None else min(size, self._dataset_size)
-        )
-        self._dataset = self._dataset.repeat().batch(self._size)
-
-        # Build string handles.
-        iterator = tf.data.make_one_shot_iterator(self._dataset)
-        self._string_handle_tensor = iterator.string_handle()
-
-        return self
-
-
-class OmniglotDataset(base.Dataset):
-    """A dataset that implements Omniglot-speicfic data preprocessing."""
+    """Provides tf.data.Dataset that represents a single Omniglot category."""
 
     RAW_IMG_SHAPE = (105, 105, 1)
     PREPROC_IMG_SHAPE = (28, 28, 1)
 
-    def __init__(self, categories, rotations=(0,), **kwargs):
-        self.rotations = tuple(rotations)
-        super(OmniglotDataset, self).__init__(categories, **kwargs)
+    def __init__(self, data_dir=None, name=None):
+        super(OmniglotCategory, self).__init__(data_dir, name)
 
-    def _get_preprocess_kwargs(self):
-        return tuple(
-            {"rotation": random.sample(self.rotations, 1)[0]} for _ in self._categories
-        )
+        # Internals.
+        self._dataset = None
 
-    @classmethod
-    def build_preprocess_kwarg_phs(cls, num_classes):
-        """Builds kwarg placeholders used for preprocessing."""
-        return tuple(
-            {
-                "rotation": tf.placeholder(
-                    dtype=tf.int32, shape=[], name=f"class_{i}_rotation"
-                )
-            }
-            for i in range(num_classes)
-        )
+    @property
+    def dataset(self) -> np.ndarray:
+        return self._dataset
 
     @classmethod
-    def preprocess(cls, example, rotation=0, **kwargs):
-        image = utils.deserialize_image(example, channels=1, shape=cls.RAW_IMG_SHAPE)
-        image = tf.image.resize(image, size=cls.PREPROC_IMG_SHAPE[:-1])
+    def preprocess(cls, image, rotation=0):
         image = tf.image.rot90(image, k=rotation)
         return image
 
+    def _build(self, max_size=None, shuffle=True):
+        # Infer dataset size.
+        filepaths = glob.glob(os.path.join(self.data_dir, "*.png"))
+        if shuffle:
+            random.shuffle(filepaths)
+        if max_size is not None:
+            filepaths = filepaths[:max_size]
+        self._size = len(filepaths)
+
+        # Load data.
+        self._dataset = []
+        for fpath in filepaths:
+            with open(fpath, "rb") as fp:
+                image = Image.open(fp).resize(self.PREPROC_IMG_SHAPE[:-1])
+                image = np.array(image).astype(np.float32)
+                self._dataset.append(np.expand_dims(image, axis=-1))
+        self._dataset = np.stack(self._dataset)
+
+        return self
+
+
+class OmniglotDataPool(base.DataPool):
+    """Manages Omniglot categories."""
+
+    def __init__(
+        self, categories: List[OmniglotCategory], name: str = "OmniglotDataPool"
+    ):
+        super(OmniglotDataPool, self).__init__(categories, name=name)
+
+    @property
+    def output_types(self):
+        return tf.float32
+
+    @property
+    def output_shapes(self):
+        return OmniglotCategory.PREPROC_IMG_SHAPE
+
+    def initialize(self, sess):
+        """Initializes all managed data categories within the given session."""
+        return self
+
+
+class OmniglotDataset(base.Dataset):
+    """Implements Omniglot-specific preprocessing functionality."""
+
+    def __init__(
+        self,
+        num_classes: int,
+        data_shapes: tf.TensorShape,
+        data_types: tf.DType,
+        rotations: Tuple[int] = (0,),
+        name: Optional[str] = None,
+    ):
+        self.num_classes = num_classes
+        self.data_shapes = data_shapes
+        self.data_types = data_types
+        self.rotations = rotations
+        self.name = name or self.__class__.__name__
+
+        # Internals.
+        self._data_tensors = None
+        self._rotations_ph = None
+
+        self.built = False
+
+    def build(self):
+        if not self.built:
+            data_tensors = []
+            with tf.name_scope(self.name):
+                # Data placeholders for each class.
+                for k in range(self.num_classes):
+                    data_ph = tf.placeholder(
+                        shape=(None,) + self.data_shapes,
+                        dtype=self.data_types,
+                        name=f"data_class_{k}",
+                    )
+                    data_tensors.append(data_ph)
+                # Rotations.
+                self._rotations_ph = tf.placeholder(
+                    tf.int32, shape=[None], name="rotations"
+                )
+            self._data_tensors = tuple(data_tensors)
+            self.built = True
+        return self
+
+    def get_preprocessor(self) -> Optional[Callable]:
+        def _inner(args):
+            image, label = args
+            image = OmniglotCategory.preprocess(
+                image, rotation=self._rotations_ph[label]
+            )
+            return image
+
+        if len(self.rotations) > 1:
+            return _inner
+
+    def get_feed_list(
+        self, data_arrays: Tuple[np.ndarray]
+    ) -> List[Tuple[tf.Tensor, Any]]:
+        feed_list = [(ph, array) for array, ph in zip(data_arrays, self._data_tensors)]
+        rotations = random.choices(self.rotations, k=self.num_classes)
+        feed_list.append((self._rotations_ph, rotations))
+        return feed_list
+
 
 class OmniglotMetaDataset(base.MetaDataset):
-    """A meta-dataset with omniglot-specific data preprocessing."""
+    """A meta-dataset that samples Omniglot datasets."""
 
     Dataset = OmniglotDataset
 
-    def __init__(
-        self, categories, num_classes, max_distinct_datasets=None, rotations=(0,)
-    ):
-        super(OmniglotMetaDataset, self).__init__(
-            categories=categories,
-            num_classes=num_classes,
-            max_distinct_datasets=max_distinct_datasets,
-        )
-        self._dataset_kwargs = {"rotations": rotations}
+    def get_feed_list_batch(
+        self, requests: Tuple[Tuple[int]]
+    ) -> List[List[Tuple[tf.Tensor, str]]]:
+        """Returns a feed list for the requested meta-batch of datasets."""
+        if len(requests) > self.batch_size:
+            raise ValueError("The dataset request is incompatible.")
+        # Get feed dicts for each request.
+        feed_lists = []
+        for n, category_ids in enumerate(requests):
+            feed_lists.append(
+                self._dataset_batch[n].get_feed_list(
+                    [self.data_pool.categories[i].dataset for i in category_ids]
+                )
+            )
+        return feed_lists
